More test refactoring
Refactor some of the test setup to make it more linear

Working on #23 and #22
