+++ servicex/servicex.py
@@ -17,12 +17,8 @@ from retry import retry
 import uproot
 
 from .utils import (
-    _default_wrapper_mgr, _run_default_wrapper, _status_update_wrapper,
-    _submit_or_lookup_transform, ServiceX_Exception, ServiceXFrontEndException)
-
-
-# Where shall we store files by default when we pull them down?
-default_file_cache_name = os.path.join(tempfile.gettempdir(), 'servicex')
+    ServiceXFrontEndException, ServiceX_Exception, _default_wrapper_mgr,
+    _run_default_wrapper, _status_update_wrapper, _submit_or_lookup_transform)
 
 
 # Number of seconds to wait between polling servicex for the status of a transform job
@@ -271,7 +267,8 @@ async def get_data_async(selection_query: str, datasets: Union[str, List[str]],
     if file_name_func is None:
         if storage_directory is None:
             def file_name(req_id: str, minio_name: str):
-                return os.path.join(default_file_cache_name, req_id,
+                import servicex.utils as sx
+                return os.path.join(sx.default_file_cache_name, req_id,
                                     santize_filename(minio_name))
             file_name_func = file_name
         else:
