+++ tests/test_servicex.py
@@ -15,20 +15,7 @@ import pytest
 
 import servicex as fe
 
-from .utils_for_testing import good_transform_request, ClientSessionMocker  # NOQA
-
-
-@pytest.fixture(autouse=True)
-def delete_default_downloaded_files():
-    download_location = os.path.join(tempfile.gettempdir(), 'servicex-testing')
-    import servicex.servicex as sx
-    sx.default_file_cache_name = download_location
-    if os.path.exists(download_location):
-        shutil.rmtree(download_location)
-    yield
-    if os.path.exists(download_location):
-        shutil.rmtree(download_location)
-
+from .utils_for_testing import good_transform_request, ClientSessionMocker, delete_default_downloaded_files  # NOQA
 
 @pytest.fixture(scope="module")
 def reduce_wait_time():
